Replace sync function with async replies
The `active_subscription_ids` function was added as a
stop-gap for the initial round of tests. However as we begin
to build out integration tests we quickly find need for more
control messages from the socket. These control messages are
now provided in the form of an Absinthe.Socket.Reply struct
that can be received by callers if they so choose.

Rather than ask uninterested developers to consume these
replies, the API has been updated to accept a `reference()`
where applicable and the presence of the reference will
determine whether or not the replies are sent.

Right now this is mostly useful in tests but it may be able
to provide a foundation for higher-level APIs as well.