Add subscribe!/1 high-level function
The `subscribe!` function works by starting a WebSocket
connection to the given url for the caller. Once the
connection has been established, then subcription messages
will be sent to the caller from the socket process.

The socket processes are supervised by the `SocketSupervisor`
process that starts as part of `AbsintheClient`'s application
supervision tree.