Set proper scheme in run_absinthe_ws_adapter/1
The WebSocket adapter now automatically sets the request URI
scheme to a valid WebSocket scheme (one of "ws" or "wss")
for HTTP schemes.

This change also ensures that the Req option `:base_url` is
taken into account when connecting to the WebSocket server.
Unfortunately this introduces lots of NotJoinedError noise in
the tests (and subsequently in user logs!) because we start
the connection and immediately attempt to push to the control
topic. We will need to address this in a future PR.