Add run/2, graphql_path/1 and other request steps
The run/2 function is the AbsintheClient version of
Req.request/1. It is a place where we can set defaults and
where we can document the options that apply to all of the
high-level client functions. In the future if we implement
query! and mutate! functions, they should ultimately
invoke run/2.

The default paths of "/graphql" and "/socket/websocket" are
automatically set for run/2 and subscribe!/2 respectively.
They can be overridden via the :url request option.