Ensure replies are not sent for re-played subscriptions
To avoid a scenario where a caller creates a subscription
synchronously, the subscription is re-played due to a
disconnect, and the caller receives an unexpected message
to its mailbox related to the re-played subscription.

In the future we may want a better way to handle control
messages like disconnect/reconnect/resubscribe, but for
now it is enough to ensure that users will not receive
unexpected messages from the WebSocket process.