Document low-level AbsintheClient.WebSocket.connect/1
Low-level WebSocket API connect/1 accepts a keyword of options
usually provided by the run_absinthe_ws step in the request
pipeline. The connect/1 function supervises all started processes
under AbsintheClient.DynamicSupervisor which is started by
AbsintheClient on app startup.

Low-level push/3 returns the push ref without wrapping it
in an :ok tuple.

High-level :web_socket option allows setting the socket to use
for a given request. This is convenient if you want to start a
socket with connect/1 and ensure that it is the socket used by
a particular request.

Additional documentation for some of the built-in timeout options.
Support for other customizing other WebSocket timeouts may be added
in the future.