Loosen restrictions on GraphQL steps
In the spirit of Req's step-driven pipeline, instead of
raising for a missing query or an unsupported HTTP method,
AbinstheClient will no-op in those cases, allowing the
end-user the opportunity to supply their own plugins to fill
in the gaps. For example, one upshot of this change is that
Ecto sandbox requests can be made through an AbsintheClient
request whereas previously the delete request would have
raised for an invalid method.

Part of the reasoning here was around WebSocket.connect and
not wanting to set a `:query` option if not necessary.
Unfortunately this did not work out quite as planned–
without setting the :query option then the proper headers are
not applied when the connection is started.