Refactor AbsintheClient API
After evaluating some of the other Req plugins in the wild I decided
that the run functions and the request steps added far too much
accidental complexity.

This is the first of two changes designed to reduce the plugin surface
area and focus on the specific functionality that the plugin adds to
the Req package.

Specifically, we can remove the pass-thru run functions and let the
user invoke the equivalent Req function of their choosing. This change
puts the user back in charge of the type of request they are performing
and allows AbsintheClient to stop being prescriptive about some choices.