Refactor WebSocket API
This refactor composes the following changes:

- Drop the subscribe!/3 pass-thru function

- Require the :web_socket option to use the WebSocket adapter

- WebSocket.connect/2 returns a Request.t()

- WebSocket.push/2 accepts graphql option from Request

In a change similar to the AbsintheClient refactor, we are no longer
providing a pass-thru function to perform a subscription operation.
Instead, users must use Req.request/2 and the :web_socket option.

To simplify the process of starting a WebSocket and using it with a
Request, WebSocket.connect/2 accepts a request and returns a request
configured to perform operations over a WebSocket.

Taken together, these changes are designed to provide more clarity
around what options are used to start the WebSocket, when the WebSocket
starts, and which WebSocket process will handle a given request.

Additionally WebSocket.push/3 is now WebSocket.push/2 and it accepts a
graphql arg in the same shape as the Request option– either a
String.t() or a {String.t(), map()} tuple.