WebSocket.connect/2 returns process name (#10)
Closes #9

Returning a Request struct with the web_socket option already
set by WebSocket.connect/2 made it difficult to re-use the same
Request for HTTP and WebSocket connections. This requires carrying
around two Request structs in the base case (the reference app) so
instead with this change we can carry around the Request and the
socket name and only apply it to operations that need the WebSocket
(i.e subscriptions).